In findBindingSite.R I modified the internal object ".nucleotides" to add: "T", "I", "DU" and "DI", which I used in the other modifications.
BUG 1:
In checkNuc and measureNuc I changed the order of some checks.
Particularly the conditional that checks whether there is a Nucleic
Acid in the structure AFTER selecting the chain. This was
generating an erro rwith the structure 1XPU.

BUG 2:
I also changed the way I identify nucleotides in the data. Instead
of using systematically the C4' atom, I changed it for an object
that I pass as argument (refatm, which I still have to document).

In pipeNucData I add a double checking for structures that only
contain phosphates, so that "refatm" is set to "P" instead of the
default "C4'". This solves the error I had with structure 1Y1Y