I found an error in the documentation: I was mentioning select.model() when I wanted to refer to selectModel(). This happened in findBindingSite() and in cifAsPDB().
Furthermore, I solved another BUG:
The structure 406D is an X-RAY with 4 different models. The models
use different ALT records. Since the pipeline (pipeNucData)
systematically selected ALT records "A" in an early stage of the
execution, the process failed when trying to read subsequent
models. To avoid this, I changed the pipeline to cope with ALT
records in a later stage of the process, when I already have
the desired model and chain.

This way I could also solve a dirty patch in the code that I had
writen to solve a similar problem for the PDB ID 3OK4.