BUG solved: Structures such as 1PBL, 1XV6 and 1DV4, which contain a Nucleic Acid, did not pass the pipeline pipeNucData() because all its bases are non-canonical.
I consider them irrelevant for the pipeline, but might be of
interest of a user at some point. Thus, I developed a new argument
(force) for measureNuc and checkNuc so that the user can force the
analysis manually for particular structures such as the ones I
mentioned before.

I included the argument in the documentation accordingly.