I found a Bug and solved it in applyToPDB() when executing:
listpdb <- c("1s72", "1bau", "1rna")
applyToPDB(queryResol, listpdb)

Because the second structure is soved by NMR, so the resolution
query returns a NULL. When the applyToPDB function tried to coerce
the result to a data.frame it completely messed up the result.

Reason:
When I unlist the output of apply and there's a NULL value inside,
it is completely removed from the output vector of unlist.