In DESCRIPTION: I added a line to compile R functions (theoreticaly faster)
In NAMESPACE:
Just changed the order of some functions to keep alphabetical order
(an error I had missed before)

In man/
I just reinstalled the packaged and the documentation was updated.

In cifAsPDB (sufcuntion)
I added a line to remove quotes from elety records when read from
cif files. I think it's not necessary anymore, but just in case...
and I removed this same line from measureNuc and checkNuc