Deveolped a data directory with a data.frame contaning some of the most expensive/common queries of the package (RNA/DNAclassifier), to avoid quering the APIs for a list of NA structures. I also modified RNAclassifier and DNAclassifier to check in the package data before sending a query.
Moreover, I apply these data also in the getAltRepres function to
improve performance (which becomes ~3 times faster).

On the other hand, I detected a problem with the classifiers for
the newest structures, since one of my queries is to the MMB API
(queryHetAtms) and the newest entries return NULL, so they are
misclassified. I will push a new commit with a solution