Added a new function queryObsoleteList, which returns an updated list of all obsolete PDB IDs. Furthermore, I also updated the queryEntryList to send the query to the RCSB API instead of ours, which usually lacks the very latest releses.
Making use of queryObsoleteList, I updated getAltRepres, which
reduced an order of magnitude its execution time (20min to 2 min
and 60s to 5s, depending on the datasets). In addition with the
previous update, the function is now fast enough (it used to take
between 40 and 60 min when querying ALL the data to the APIs).