Updated the data in the package (fastquery and entities). Updated the way classifyNA functions acces to it. Also added the fastquery feature to the queryTechnique and queryResol functions. Today I found a new type of entity molecule-type: cyclic-pseudo-peptide. I added in the group of proteins (L). I also modified the countEntites function to be able to access the presaved data in entities.
Finally, I generated a first test function for the fastquery data,
but I havent tested it and might never do it, since it will
produce thousands of queries to the APIs. I might even remove this
test before releasing the package, but for the moment it showed me
how to do test functions.